Vulkan: Split off and implement SwapChain
